Fix app V1

+++ AccountingSystem/AccountingSystem.py
@@ -3,22 +3,21 @@ import pymysql
 import json
 import datetime
 
-# 一个样例
-def demo():
 
-    acc = Account("Administrator", 1)
+def demo():  # 一个样例
+    acc = Account()
     # print(acc.getitem(1))
     # acc.sellgoods(1, 2)
     print(acc.getbalance())
     print("finished")
     
-# 数据库连接
-class Sql:
-    connect = pymysql.connect("localhost", "root", "root", "AccountingSystem") # host, user, password, database
+
+class Sql:  # 数据库连接
+    connect = pymysql.connect("localhost", "root", "jiaxing+", "AccountingSystem")  # host, user, password, database
     cur = connect.cursor()
 
-# 用户
-class User:
+
+class User:  # 用户
 
     def __init__(self, id, name, password, kind):
         self.id = id                # 序号
@@ -26,11 +25,11 @@ class User:
         self.password = password    # 密码
         self.kind = kind            # 类型
 
-    def tojson(self):
-        return json.dumps([{"id" : self.id, "name" : self.name, "password" : self.password, "kind" : self.kind}])
+    def todict(self):
+        return {"id": self.id, "name": self.name, "password": self.password, "kind": self.kind}
 
-# 货物
-class Item:
+
+class Item:  # 货物
 
     def __init__(self, id, name, price, amount):
         self.id = id                # 序号
@@ -38,11 +37,11 @@ class Item:
         self.price = price          # 单价
         self.amount = amount        # 库存
 
-    def tojson(self):
-        return json.dumps([{"id" : self.id, "name" : self.name, "price" : self.price, "amount" : self.amount}])
+    def todict(self):
+        return {"id": self.id, "name": self.name, "price": self.price, "amount": self.amount}
 
-# 交易记录
-class Transaction:
+
+class Transaction:  # 交易记录
 
     def __init__(self, id, kind, itemID, amount, totalPrice, userID, time):
         self.id = id                     # 序号
@@ -52,72 +51,86 @@ class Transaction:
         self.totalPrice = totalPrice     # 交易额
         self.userID = userID             # 用户序号
         self.time = time                 # 时间
-    def tojson(self):
-        return json.dumps([{"id" : self.id, "kind" : self.kind, "itemID" : self.itemID, "amount" : self.amount, "totalPrice" : self.totalPrice, "userID" : self.userID, "time" : self.time}])
 
-# 收支
-class Balance:
+    def todict(self):
+        return {"id": self.id, "kind": self.kind, "itemID": self.itemID, "amount": self.amount, "totalPrice": self.totalPrice, "userID": self.userID, "time": self.time}
+
+
+class Balance:  # 收支
 
     def __init__(self, id, profit, date):
         self.id = id            # 序号
         self.profit = profit    # 盈亏
         self.date = date        # 日期
 
-    def tojson(self):
-        return json.dumps([{"date" : self.date, "sum" : self.profit}])
+    def todict(self):
+        return {"date": self.date, "sum": self.profit}
 
-# 账户
-class Account:
 
-    def __init__(self, kind, userID):
-        self.kind = kind        # Administrator, Seller, Buyer, Accountant
-        self.userID = userID
+class Account:  # 账户
 
-    def getitem(self, id):
-        Sql.cur.execute("SELECT * FROM Item WHERE ID = '%s'" % id)
+    def __init__(self):
+        self.kind = ''        # Administrator, Seller, Buyer, Accountant
+        self.userID = 0
+        self.state = False
+
+    def getitem(self, name):
+        if self.state:
+            raise RuntimeError
+        Sql.cur.execute("SELECT * FROM Item WHERE Name = '%s'" % name)
         data = Sql.cur.fetchone()
         item = Item(data[0], data[1], data[2], data[3])
-        return item.tojson()
+        return item.todict()
 
-    def sellgoods(self, id, amount):
+    def sellgoods(self, name, amount):
+        if self.state:
+            raise RuntimeError
         if (self.kind != "Administrator") and (self.kind != "Seller"): raise RuntimeError("Permission denied.")
-        Sql.cur.execute("SELECT * FROM Item Where ID = '%s'" % id)
+        Sql.cur.execute("SELECT * FROM Item Where Name = '%s'" % name)
         data = Sql.cur.fetchone()
         oldamount = data[3]
         if (oldamount < amount): raise RuntimeError("Insufficient goods.")
-        Sql.cur.execute("UPDATE Item SET Amount = %f Where ID = '%s'" % (oldamount - amount, id))
+        Sql.cur.execute("UPDATE Item SET Amount = %f Where Name = '%s'" % (oldamount - amount, name))
         Sql.connect.commit()
         price = data[2]
         date = datetime.datetime.now().strftime("%Y-%m-%d")
-        Sql.cur.execute("INSERT INTO Transaction VALUES(%d, '%s', %d, %d, %f, %d, '%s')" % (self.nextid("Transaction"), "Sell", id, amount, price * amount, self.userID, date))
+        Sql.cur.execute("INSERT INTO Transaction VALUES(%d, '%s', %d, %d, %f, %d, '%s')" % (self.nextid("Transaction"), "Sell", name, amount, price * amount, self.userID, date))
         Sql.connect.commit()
         
-    def buygoods(self, id, amount):
+    def buygoods(self, name, amount):
+        if self.state:
+            raise RuntimeError
         if (self.kind != "Administrator") and (self.kind != "Buyer"): raise RuntimeError("Permission denied.")
-        Sql.cur.execute("SELECT * FROM Item Where ID = '%s'" % id)
+        Sql.cur.execute("SELECT * FROM Item Where Name = '%s'" % name)
         data = Sql.cur.fetchone()
         oldamount = data[3]
-        Sql.cur.execute("UPDATE Item SET Amount = %f Where ID = '%s'" % (oldamount + amount, id))
+        Sql.cur.execute("UPDATE Item SET Amount = %f Where Name = '%s'" % (oldamount + amount, name))
         Sql.connect.commit()
         price = data[2]
         date = datetime.datetime.now().strftime("%Y-%m-%d")
-        Sql.cur.execute("INSERT INTO Transaction VALUES(%d, '%s', %d, %d, %f, %d, '%s')" % (self.nextid("Transaction"), "Buy", id, amount, - price * amount, self.userID, date))
+        Sql.cur.execute("INSERT INTO Transaction VALUES(%d, '%s', %d, %d, %f, %d, '%s')" % (self.nextid("Transaction"), "Buy", name, amount, - price * amount, self.userID, date))
         Sql.connect.commit()
 
     def additem(self, name, price):
+        if self.state:
+            raise RuntimeError
         if (self.kind != "Administrator") and (self.kind != "Buyer"): raise RuntimeError("Permission denied.")
         Sql.cur.execute("INSERT INTO Item VALUES(%d, '%s', %f, %d)" % (self.nextid("Item"), name, price, 0))
         Sql.connect.commit()
     
     def getitemlist(self):
+        if self.state:
+            raise RuntimeError
         Sql.cur.execute("SELECT * FROM Item")
         res = []
         for data in Sql.cur.fetchall():
             item = Item(data[0], data[1], data[2], data[3])
-            res.append(item.tojson())
+            res.append(item.todict())
         return res
 
     def getbalance(self):
+        if self.state:
+            raise RuntimeError
         if (self.kind != "Administrator") and (self.kind != "Accountant"): raise RuntimeError("Permission denied.")
         Sql.cur.execute("SELECT * FROM Transaction")
         profit = 0.0
@@ -129,23 +142,29 @@ class Account:
                 else:
                     profit += data[4]
         balance = Balance(self.nextid("Balance"), profit, date)
-        return balance.tojson()
+        return balance.todict()
 
     def getuserlist(self):
+        if self.state:
+            raise RuntimeError
         if (self.kind != "Administrator"): raise RuntimeError("Permission denied.")
         Sql.cur.execute("SELECT * FROM User")
         res = []
         for data in Sql.cur.fetchall():
             item = Item(data[0], data[1], data[2], data[3])
-            res.append(item.tojson())
+            res.append(item.todict())
         return res
 
     def removeuser(self, id):
+        if self.state:
+            raise RuntimeError
         if (self.kind != "Administrator"): raise RuntimeError("Permission denied.")
         Sql.cur.execute("DELETE FROM User WHERE ID = %d" % id)
         Sql.connect.commit()
 
     def adduser(self, name, password, kind):
+        if not self.state:
+            raise RuntimeError
         if (self.kind != "Administrator"): raise RuntimeError("Permission denied.")
         Sql.cur.execute("INSERT INTO User VALUES(%d, '%s', '%s', '%s')" % (self.nextid("User"), name, password, kind))
         Sql.connect.commit()
@@ -154,10 +173,24 @@ class Account:
         Sql.cur.execute("SELECT MAX(ID) FROM %s" % table)
         data = Sql.cur.fetchone()
         res = data[0]
-        if res == None:
+        if res is None:
             return 1
         else:
             return res + 1
 
-if __name__=="__main__":
+    def login(self, user, password):
+        Sql.cur.execute("SELECT * FROM User WHERE Name = '%s' AND Password = '%s'" % (user, password))
+        data = Sql.cur.fetchone()
+        if data:
+            self.kind = data[3]
+            self.userID = data[0]
+            self.state = True
+        else:
+            raise RuntimeError("Wrong username or password.")
+
+    def logout(self):
+        self.state = False
+
+
+if __name__ == "__main__":
     demo()
